pyproject.toml
[project]
name = "dataeng"
version = "0.1.0"
description = ""
authors = [
    {name = "Teemu Jäppinen",email = "teemu.t.jappinen@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "mysql-connector-python (>=9.2.0,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "dataeng.main:start"
